+++ R/plot.corrected_lasso.R
@@ -1,7 +1,8 @@
 #' @title plot.corrected_lasso
-#' @description Plot the output of fit_corrected_lasso
-#' @param x Object of class corrected_lasso, returned from calling fit_corrected_lasso()
-#' @param type Type of plot. Either "nonzero" or "path".
+#' @description Plot the output of corrected_lasso
+#' @param x Object of class corrected_lasso, returned from calling corrected_lasso()
+#' @param type Type of plot. Either "nonzero" or "path". Ignored if \code{length(x$radii) == 1},
+#' in case of which all coefficient estimates are plotted at the given regularization parameter.
 #' @param ... Other arguments to plot (not used)
 #' @examples
 #' # Example with linear regression
@@ -19,13 +20,21 @@
 #' # Response
 #' y <- X %*% beta + rnorm(n, sd = 1)
 #' # Run the corrected lasso
-#' fit <- fit_corrected_lasso(W, y, sigmaUU, family = "gaussian")
+#' fit <- corrected_lasso(W, y, sigmaUU, family = "gaussian")
 #' plot(fit)
 #'
 #' @export
 plot.corrected_lasso <- function(x, type = "nonzero", ...) {
 
-  if(type == "nonzero") {
+  if(length(x$radii) == 1){
+    message("Only one regularization parameter. Plotting all coefficients.\n")
+    df <- data.frame(
+      coefficient = 1:length(x$betaCorr),
+      estimate = x$betaCorr
+    )
+    ggplot2::ggplot(df, ggplot2::aes_(x =~ coefficient, y =~ estimate)) +
+      ggplot2::geom_point()
+  } else if(type == "nonzero") {
     df <- data.frame(radius = x$radii, nonZero = colSums(abs(x$betaCorr) > 0))
     ggplot2::ggplot(df, ggplot2::aes_(x =~ radius, y =~ nonZero)) +
       ggplot2::geom_line() +